fix(mobile): ChunkedUploader — AtomicLong for progress, Long for file size
- Replace synchronized(coroutineContext) with AtomicLong for thread-safe
  progress tracking (previous monitor was unreliable)
- Change file.length().toInt() to Long to support files >2GB
- Change UploadsApi.init() totalSize param from Int to Long
- Update readFileChunk() params from Int to Long
- Add UploadWorkerTest, UploadSchedulerTest, extended ChunkedUploaderTest
- Add UploadsApiTest for >2GB file size (no Int overflow)

+++ mobile/shared/src/commonMain/kotlin/ru/skatelab/shared/api/UploadsApi.kt
@@ -11,7 +11,7 @@ import kotlinx.serialization.Serializable
 
 class UploadsApi(private val client: HttpClient) {
     /** Initialize a multipart upload. Backend uses query parameters. */
-    suspend fun init(fileName: String, contentType: String, totalSize: Int): UploadInitResponse =
+    suspend fun init(fileName: String, contentType: String, totalSize: Long): UploadInitResponse =
         client.post("/uploads/init") {
             parameter("file_name", fileName)
             parameter("content_type", contentType)
